feat: 실제 swagger에 맞춰서 사용할 수 있게끔 변경하고 페이징 처리 추가

+++ hr_bank/src/main/java/com/codeit/hrbank/repository/BackupQueryRepository.java
@@ -13,7 +13,7 @@ public interface BackupQueryRepository {
       BackupStatus status,
       Instant startedAtFrom,
       Instant startedAtTo,
-      int idAfter,
+      Long idAfter,
       String cursor,
       int size,
       String sortField,
